+++ src/components/tables/PoolsTable/PoolsTable.vue
@@ -235,7 +235,7 @@ function navigateToPoolMigration(pool: DecoratedPoolWithShares) {
 
 function getTotalRewardsAPR(pool: DecoratedPoolWithShares) {
   return bnum(pool?.dynamic?.apr?.staking?.Rewards || '0').plus(
-    pool?.dynamic?.apr?.total
+    pool?.dynamic?.apr?.total || '0'
   );
 }
 </script>
