+++ src/components/tooltips/LiquidityAPRTooltip.vue
@@ -42,7 +42,7 @@ const lmBreakdown = computed(
 );
 
 const validAPR = computed(
-  () => Number(props.pool?.dynamic?.apr?.total) * 100 <= APR_THRESHOLD
+  () => Number(props.pool?.dynamic?.apr?.total || '0') * 100 <= APR_THRESHOLD
 );
 
 const lmTokens = computed(() => getTokens(Object.keys(lmBreakdown.value)));
@@ -52,7 +52,7 @@ const lmMultiRewardPool = computed(
 );
 
 const hasThirdPartyAPR = computed(() =>
-  bnum(props.pool?.dynamic?.apr?.thirdParty).gt(0)
+  bnum(props.pool?.dynamic?.apr?.thirdParty || '0').gt(0)
 );
 
 const thirdPartyBreakdown = computed(
